comments issues solved
