+++ modules/displayPokemon.js
@@ -4,9 +4,6 @@ import displayMessage from './displayMessage.js';
 const cardPoke = document.querySelector('.cardModal');
 const $grafica = document.querySelector('#grafica');
 
-const nameInput = document.querySelector('.name-form');
-const messageInput = document.querySelector('.insights-form');
-
 const displayPokemon = async (card, index) => {
   console.log(`imprimiendo poke ${+card[index].data.id}`);
   let nroAbility = 0;
@@ -77,26 +74,6 @@ const displayPokemon = async (card, index) => {
   const a = await getComments(card[index].data.id);
   console.log(`estos son los comentarios de ${card[index].data.id}:\n${JSON.stringify(a)}`);
   displayMessage(a);
-  // eslint-disable-next-line no-const-assign
-  document.querySelector('form').addEventListener('submit', async (e) => {
-    e.preventDefault();
-    // eslint-disable-next-line no-empty
-    if (nameInput.value === '' || messageInput === '') {} else {
-      console.log(`añadiendo comentario al ${card[index].data.id}`);
-      await postComment(card[index].data.id, nameInput.value, messageInput.value);
-      console.log(`comentario del ${card[index].data.id}`, card[index].data.id, nameInput.value, messageInput.value);
-      console.log(`obteniendo comentarios del ${card[index].data.id}`);
-      const b = await getComments(card[index].data.id);
-      console.log(`estos son los comentarios de ${card[index].data.id}:\n${JSON.stringify(b)}`);
-      // console.log('its:', b);
-      // console.log('its me');
-      displayMessage(b);
-      // eslint-disable-next-line no-const-assign
-      nameInput.value = '';
-      // eslint-disable-next-line no-const-assign
-      messageInput.value = '';
-    }
-  });
 };
 
 export default displayPokemon;
